refactor: make MyselfHandler use KnitterService
