+++ src/main/kotlin/com/kroffle/knitting/infra/persistence/design/repository/R2dbcDesignRepository.kt
@@ -4,7 +4,7 @@ import com.kroffle.knitting.domain.design.entity.Design
 import com.kroffle.knitting.infra.persistence.design.entity.DesignEntity
 import com.kroffle.knitting.infra.persistence.design.entity.toDesignEntity
 import com.kroffle.knitting.infra.persistence.helper.pagination.PaginationHelper
-import com.kroffle.knitting.usecase.design.DesignRepository
+import com.kroffle.knitting.usecase.repository.DesignRepository
 import com.kroffle.knitting.usecase.helper.pagination.type.Paging
 import com.kroffle.knitting.usecase.helper.pagination.type.Sort
 import com.kroffle.knitting.usecase.helper.pagination.type.SortDirection
