+++ src/main/kotlin/com/kroffle/knitting/infra/persistence/knitter/repository/R2dbcKnitterRepository.kt
@@ -2,7 +2,7 @@ package com.kroffle.knitting.infra.persistence.knitter.repository
 
 import com.kroffle.knitting.domain.knitter.entity.Knitter
 import com.kroffle.knitting.infra.persistence.knitter.entity.toKnitterEntity
-import com.kroffle.knitting.usecase.auth.KnitterRepository
+import com.kroffle.knitting.usecase.repository.KnitterRepository
 import reactor.core.publisher.Mono
 
 class R2dbcKnitterRepository(private val repository: DBKnitterRepository) : KnitterRepository {
