--- src/main/kotlin/com/kroffle/knitting/usecase/auth/KnitterRepository.kt
@@ -1,3 +1,0 @@
-package com.kroffle.knitting.usecase.auth
-
-interface KnitterRepository : AuthService.KnitterRepository
