--- src/main/kotlin/com/kroffle/knitting/usecase/design/DesignRepository.kt
@@ -1,3 +1,0 @@
-package com.kroffle.knitting.usecase.design
-
-interface DesignRepository : DesignService.DesignRepository
