--- src/main/kotlin/com/kroffle/knitting/usecase/product/ProductRepository.kt
@@ -1,3 +1,0 @@
-package com.kroffle.knitting.usecase.product
-
-interface ProductRepository : ProductService.ProductRepository
