+++ src/main/kotlin/com/kroffle/knitting/usecase/repository/DesignRepository.kt
@@ -1,0 +1,5 @@
+package com.kroffle.knitting.usecase.repository
+
+import com.kroffle.knitting.usecase.design.DesignService
+
+interface DesignRepository : DesignService.DesignRepository
