+++ src/main/kotlin/com/kroffle/knitting/usecase/repository/ProductRepository.kt
@@ -1,0 +1,5 @@
+package com.kroffle.knitting.usecase.repository
+
+import com.kroffle.knitting.usecase.product.ProductService
+
+interface ProductRepository : ProductService.ProductRepository
