+++ src/test/kotlin/com/kroffle/knitting/controller/router/auth/LoginRouterTest.kt
@@ -15,7 +15,7 @@ import com.kroffle.knitting.infra.oauth.dto.GoogleOAuthConfig
 import com.kroffle.knitting.infra.persistence.knitter.entity.KnitterEntity
 import com.kroffle.knitting.infra.properties.WebApplicationProperties
 import com.kroffle.knitting.usecase.auth.AuthService
-import com.kroffle.knitting.usecase.auth.KnitterRepository
+import com.kroffle.knitting.usecase.repository.KnitterRepository
 import com.kroffle.knitting.usecase.auth.dto.OAuthProfile
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
