+++ src/test/kotlin/com/kroffle/knitting/controller/router/product/ProductRouterTest.kt
@@ -20,7 +20,7 @@ import com.kroffle.knitting.helper.extension.addDefaultRequestHeader
 import com.kroffle.knitting.helper.extension.like
 import com.kroffle.knitting.infra.jwt.TokenDecoder
 import com.kroffle.knitting.infra.properties.WebApplicationProperties
-import com.kroffle.knitting.usecase.product.ProductRepository
+import com.kroffle.knitting.usecase.repository.ProductRepository
 import com.kroffle.knitting.usecase.product.ProductService
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.BeforeEach
